refactor: expose jmespath powertools functions
